created videoPlayer

+++ app/src/main/java/com/nroncari/movieplay/data/model/MovieDetailResponse.kt
@@ -3,7 +3,7 @@ package com.nroncari.movieplay.data.model
 import com.google.gson.annotations.SerializedName
 
 data class MovieDetailResponse(
-    @SerializedName("id") val id: Int? = 0,
+    @SerializedName("id") val id: Long? = 0,
     @SerializedName("original_title") val originalTitle: String? = null,
     @SerializedName("overview") val overview: String? = null,
     @SerializedName("poster_path") val posterPath: String? = null,
